+++ packages/terra-popup/src/terra-dev-site/test/popup/DifferentAttachmentsPopup.test.jsx
@@ -82,6 +82,7 @@ class AlignmentPopup extends React.Component {
             className={cx('popup-button')}
             onClick={this.handleButtonClick}
             ref={this.setButtonNode}
+            aria-label="Alignment button"
           >
             {''}
           </button>