+++ packages/terra-aggregator/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to terra-functional-testing
+
 ## 4.65.0 - (February 9, 2021)
 
 * Changed
@@ -20,7 +23,7 @@
 ## 4.62.0 - (January 5, 2021)
 
 * Fixed
- * Fixed broken links in documentation.
+  * Fixed broken links in documentation.
 
 ## 4.61.0 - (December 8, 2020)