+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to use terra-functional-testing
+
 ## 4.10.0 - (February 9, 2021)
 
 * Changed
@@ -20,7 +23,7 @@
 ## 4.7.0 - (January 5, 2021)
 
 * Fixed
- * Fixed broken links in documentation.
+  * Fixed broken links in documentation.
 
 ## 4.6.0 - (December 8, 2020)